Initial commit of Mist-to-OMI gateway.
Basic test functionality:

The program expects to find a core named "Jan" on the wld, and a friend
request is sent to that core.

When a peer becomes available, the program starts a mist.control.follow
on the peer and if the endpoint named 'relay' changes, the new value
will be written on the OmiNode to an info item of with the same name.